refactoring - move now obselete functions from cbird into legacy. hope to delete this later
